RollTheDice/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
django-bootstrap-v5==1.0.11
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
Flask==3.0.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
django-cors-headers==4.3.1
Django==4.2.3
gunicorn==21.2.0
psycopg2-binary==2.9.6
flake8==7.0.0